Moved mock kwik file creation to io.mock.kwik.
